\name{BYMScale-package}
\alias{BYMScale-package}
\alias{BYMScale}
\docType{package}
\title{
\packageTitle{BYMScale}
}
\description{
\packageDescription{BYMScale}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BYMScale}
\packageIndices{BYMScale}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BYMScale}

Maintainer: \packageMaintainer{BYMScale}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
